arm_mpam: resctrl: Allow resctrl to allocate monitors
commit 0dce3500dc7d5f5fa36b0c90cf17d692db1e2bb2 from
git@git.assembla.com:cavium/WindRiver.linux.git

When resctrl wants to read a domain's 'QOS_L3_OCCUP', it needs
to allocate a monitor on the corresponding resource. Monitors are
allocated by class instead of component.

Add helpers to do this.

The MBM events depend on having their monitors allocated at init
time so that they can be left running. The value USE_RMID_IDX
is out of range for a monitor, and is used to indicate this
behaviour.

resctrl_arch_mon_ctx_alloc() is implemented to have a no_wait
version and a waitqueue for callers that sleep. The no_wait
version will later become an interface for the resctrl_pmu
to use.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>