untested: arm_mpam: resctrl: Add support for mbm counters
commit 062ceed48cd20f3c9f97dccde1a83d38873cb0a0 from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl has two types of counters, NUMA-local and global. MPAM has only
bandwidth counters, but the position of the MSC may mean it counts
NUMA-local, or global traffic.

But the topology information is not available.

Apply a hueristic: the L2 or L3 supports bandwidth monitors, these are
probably NUMA-local. If the memory controller supports bandwidth
monitors, they are probably global.

This selection is made from mpam_resctrl_resource_init(), which
implies resources that can be used for resctrl controls exist and
also have counters. This would be a problem on a platform that only
supports monitoring.

TODO: Add an extra pass of all classes to find the classes to use
as bandwidth counters.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>