arm_mpam: Add helper to reset saved mbwu state
commit e9b4ed4fa838d32e0ab4dc224c5e597a4ace21fe from
git@git.assembla.com:cavium/WindRiver.linux.git

resctrl expects to reset the bandwidth counters when the filesystem
is mounted.

To allow this, add a helper that clears the saved mbwu state. Instead
of cross calling to each CPU that can access the component MSC to
write to the counter, set a flag that causes it to be zero'd on the
the next read. This is easily done by forcing a configuration update.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>