octeontx2-af: Hardware configuration for inline IPsec
commit 4b5a3ab17c6c942bd428984b6b37fe3c07f18ab3 upstream

On OcteonTX2/CN10K SoC, the admin function (AF) is the only one
with all priviliges to configure HW and alloc resources, PFs and
it's VFs have to request AF via mailbox for all their needs.
This patch adds new mailbox messages for CPT PFs and VFs to configure
HW resources for inline-IPsec.

Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Signed-off-by: Srujana Challa <schalla@marvell.com>
Signed-off-by: Vidya Sagar Velumuri <vvelumuri@marvell.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
Signed-off-by: Rui He <rui.he@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>