iommu: Add helpers to retrieve iommu_groups by id or kobject
commit 03bb2b17994f709f13c437e4891d4cc4a24f3a95 from
git@git.assembla.com:cavium/WindRiver.linux.git

ARM SMMU with MPAM support are able to mark streams of traffic with
the QoS labels MPAM uses. The user-space interface for MPAM is the
resctrl filesystem, which allows threads to be moved between groups,
its natural to do the same for iommu_groups.

The resctrl interface lists threads, so will also need to list
iommu_groups, it will be necessary to walk the list of iommu_groups.
To ensure this matches what user-space sees via sysfs, it is best
to walk the kobjects.

When making a change, resctrl will only have the id of a group.
To avoid walking the list of kobjects in this case, add
iommu_group_get_by_id().

Signed-off-by: James Morse <james.morse@arm.com>
Change-Id: I58e8d9620ef72f2f57a1ad87728adcaf48ba8ec2
Signed-off-by: Siyu Zhang <siyu.zhang@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>