iommu/arm-smmu-v3: Register SMMU capabilities with MPAM
commit 849a285669c2d84318a6f21cd1dedac7eea360cb from
git@git.assembla.com:cavium/WindRiver.linux.git

Traffic in the system can be tagged with a PARTID and PMG. Different
requestors can support a different number of bits for these fields.

Before MPAM can be used, the MPAM driver has to discover the minimum
number of bits supported by any requestor, which affects the range
of PARTID and PMG that can be used.

Detect whether the SMMU supports MPAM, if it does provide the MPAM
driver with the maximum PARTID and PMG values.

Tested-by: Amit Singh Tomar <amitsinght@marvell.com>
Signed-off-by: James Morse <james.morse@arm.com>
Change-Id: Id299aa831c740976dbb5fd1b8aff346b24522dc4
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>