octeontx2-pf: replace bitmap_weight with bitmap_empty where appropriate
commit 2b330a6a6c21a9b9f4ce71e5c0be494ab577423c upstream

In some places, octeontx2 code calls bitmap_weight() to check if any bit of
a given bitmap is set. It's better to use bitmap_empty() in that case
because bitmap_empty() stops traversing the bitmap as soon as it finds
first set bit, while bitmap_weight() counts all bits unconditionally.

Signed-off-by: Yury Norov <yury.norov@gmail.com>
Integrated-by: Rui He <rui.he@windriver.com>