kernfs: move struct kernfs_root out of the public view.
commit f2eb478f2f322217aa642e11c1cc011f99c797e6 upstream

There is no need to have struct kernfs_root be part of kernfs.h for
the whole kernel to see and poke around it.  Move it internal to kernfs
code and provide a helper function, kernfs_root_to_node(), to handle the
one field that kernfs users were directly accessing from the structure.

Cc: Imran Khan <imran.f.khan@oracle.com>
Acked-by: Tejun Heo <tj@kernel.org>
Link: https://lore.kernel.org/r/20220222070713.3517679-1-gregkh@linuxfoundation.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: Siyu Zhang <siyu.zhang@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>