eth: nfp: replace driver's "pf" lock with devlink instance lock
commit 162cca42920c2597298fcdf85538b40356c92151 upstream

The whole reason for existence of the pf mutex is that we could
not lock the devlink instance around port splitting. There are
more types of reconfig which can make ports appear or disappear.
Now that the devlink instance lock is exposed to drivers and
"locked" helpers exist we can switch to using the devlink lock
directly.

Next patches will move the locking inside .port_(un)split to
the core.

Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Rui He <rui.he@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>