devlink: pass devlink_port to port_split / port_unsplit callbacks
commit 706217c1ceb516c96283a1557a31fe003d0c8052 upstream

Now that devlink ports are protected by the instance lock
it seems natural to pass devlink_port as an argument to
the port_split / port_unsplit callbacks.

This should save the drivers from doing a lookup.

In theory drivers may have supported unsplitting ports
which were not registered prior to this change.

Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Tested-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Integrated-by: Rui He <rui.he@windriver.com>