mmc: improve API to make clear hw_reset callback is for cards
commit 32f18e596141f40dbc5d8700b2730371ffd3055f upstream

To make it unambiguous that the hw_reset callback is for cards and not
for controllers, we add 'card' to the callback name and convert all
users in one go. We keep the argument as mmc_host, though, because the
callback is used very early when mmc_card is not yet populated.

Signed-off-by: Wolfram Sang <wsa+renesas@sang-engineering.com>
Link: https://lore.kernel.org/r/20220408080045.6497-4-wsa+renesas@sang-engineering.com
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
Signed-off-by: Jinchao Li <jinchao.li@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>