octeontx2-bphy-netdev: cnf95n: fix ptp timestamp in master mode
commit d46680f5f497a538e7889520675f6b2b520be350 from
git@git.assembla.com:cavium/WindRiver.linux.git

PTP master support was broken as a result of recent code changes to
support PTP slave on CNF95xxN. The reason for this is that the PTP
timestamp value includes the PHC SW timecounter offset in master mode
which is not intended. Originally, in PTP master mode, the final PTP
timestamp is computed from the BCN offset, which will be initialized
to GPS epoch. This patch fixes the problem.

Change-Id: I0070e9e5524e64bf9b8296a556e06b77d03d32b7
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/81790
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Integrated-by: Rui He <rui.he@windriver.com>