octeontx2-af: add support for SSO WQE stashing
commit 5f431c97c3055c38c80f70a8035629d511829a7f from
git@git.assembla.com:cavium/WindRiver.linux.git

SSO supports stashing WQE to L2 caches of the core its scheduled on.
To enable stashing we need to configure the offset and range of the
stash described in number of cachelines.
Add mbox to enable/disable stashing, set the stashing range, offset
on a specific HWGRP.

Signed-off-by: Pavan Nikhilesh <pbhagavatula@marvell.com>
Change-Id: I39e11076fc8128e3ed8a42943c548835fe29ca64
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/83059
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Sunil Goutham <sgoutham@marvell.com>
Integrated-by: Rui He <rui.he@windriver.com>