octeontx2-bphy-netdev:: Share RFOE PTP offset with host
commit ec9f599a538c539548d11c72bd2b1bdc38cd6da8 from
git@git.assembla.com:cavium/WindRiver.linux.git

Share the offset that is added to the PTP counter to get the
UTC time with the host.  This enables the host to use this offset
along with reading the PTP counter to directly use the Octeon EP
PTP counter as a timesource.  The offset is stored in
the unused MIO_PTP_CKOUT_THRESH_HI register.
Note that this offset is not directly maintained in the driver,
but calculated at initialization time, then updated as needed.

IPBUSW-19810
Change-Id: Iaececc6327d9a748dcad6d7366d78ebf34dfe047
Signed-off-by: Roy Franz <rfranz@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/85807
Reviewed-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-by: Satananda Burla <sburla@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Integrated-by: Rui He <rui.he@windriver.com>