sch_htb: Allow HTB quantum parameter in offload mode
commit 9fe63d5f1da939855bdfaebfd9e95c96938b6411 upstream

The current implementation of HTB offload returns the EINVAL error for
quantum parameter. This patch removes the error returning checks for
'quantum' parameter and populates its value to tc_htb_qopt_offload
structure such that driver can use the same.

Add quantum parameter check in mlx5 driver, as mlx5 devices are not capable
of supporting the quantum parameter when htb offload is used. Report error
if quantum parameter is set to a non-default value.

Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Signed-off-by: Hariprasad Kelam <hkelam@marvell.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
[Partial: not porting file drivers/net/ethernet/mellanox/mlx5/core/en/qos.c]
Signed-off-by: Rui He <rui.he@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>