iommu: Add helpers to get and set the QoS state
commit e5623274d448ad0f419d505e8deaabf802a5c5c3 from
git@git.assembla.com:cavium/WindRiver.linux.git

To allow an iommu_group to be moved between resctrl groups as if it
were a CPU thread, the mpam driver needs to be able to set the partid
and pmg for the iommu_group.

Add helpers that call the iommu driver's get/set methods for these
parameters.

Signed-off-by: James Morse <james.morse@arm.com>
Change-Id: I6d6b8d93a0d9d6fdad40ed37dc7136e09ca4d478
Signed-off-by: Siyu Zhang <siyu.zhang@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>