llist: add interface to check if a node is on a list.
commit 0ac5600375f762afa14cb8fc06cd26b3e331ec64 from
git@git.assembla.com:cavium/WindRiver.linux.git

With list.h lists, it is easy to test if a node is on a list, providing
it was initialised and that it is removed with list_del_init().

This patch provides similar functionality for llist.h lists.

 init_llist_node()
marks a node as being not-on-any-list be setting the ->next pointer to
the node itself.
 llist_on_list()
tests if the node is on any list.
 llist_del_first_init()
remove the first element from a llist, and marks it as being off-list.

Signed-off-by: NeilBrown <neilb@suse.de>
Signed-off-by: Chuck Lever <chuck.lever@oracle.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>