ACPI / PPTT: Add a helper to fill a cpumask from a processor container
commit 93946f0b359b717afec4eea63ee0a9fc437cf7b5 from
git@git.assembla.com:cavium/WindRiver.linux.git

The ACPI table for MPAM describes a set of CPUs with the UID of a
processor container. These exist both in the namespace and the PPTT.

Using the existing for-each helpers, provide a helper to find the
specified processor container in the PPTT, and fill a cpumask
with the CPUs that belong to it.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Jinchao Li <jinchao.li@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>