arm_mpam: resctrl: Determine if any exposed counter can overflow
commit 2a0a020c8ac59d0d28f99d738f42cc7f15ada163 from
git@git.assembla.com:cavium/WindRiver.linux.git

Resctrl has an overflow handler that runs on each domain every second
to ensure that any overflow of the hardware counter is accounted for.

MPAM can have counters as large as 63 bits, in which case there is no
need to check for overflow.

To allow the overflow hander to be disabled, determine if an overflow
can happen. If a class is not implemented, or has the 63bit counter,
it can't overflow.

Signed-off-by: James Morse <james.morse@arm.com>
Integrated-by: Siyu Zhang <siyu.zhang@windriver.com>