drivers:mmc:cadence: Calculate delay-element
commit 08c6f18dd199955151c0db752d3edef13a6dd54d from
git@git.assembla.com:cavium/WindRiver.linux.git

The Cadence SDHCI driver uses a delay element count when the DLL
does not lock.  This patch allows the driver to automatically
calculate the number of delay elements to use to meet the timing
rather than use hard-coded values.  This allows for variation between
chips.  The delay element is not used when the DLL is locked.

Change-Id: I7d8c27612831bf981709ce702061f7924fe5cae7
Signed-off-by: Aaron Williams <awilliams@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/124215
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Integrated-by: Jinchao Li <jinchao.li@windriver.com>