fs/resctrl: Remove unnecessary includes
commit bf7cb3f7c2a28ef837777561f889edb86fe0de63 from
git@git.assembla.com:cavium/WindRiver.linux.git

When splitting and moving the resctrl code to live in fs/resctrl and
arch/x86, some code was duplicated. This was done to keep the parser
in the script that does the moving simple. These extra includes are
harmless on x86.

Remove them to allow other architectures to start using fs/resctrl.

Signed-off-by: James Morse <james.morse@arm.com>
Signed-off-by: Siyu Zhang <siyu.zhang@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>