octeontx2-af: cn20k: cpt: Use proper mask to get max CPT LF's
commit 196548896b1596a4bed2f881911bca403dca02ec from
git@git.assembla.com:cavium/WindRiver.linux.git

The number of bits implemented to determine the total
number of CPT LF's differs between 10x, and 20x. For
instance, on 10x bits 7:0 (of CPTX_AF_CONSTANTS0) are
used to determine number of LF's whereas on 20x, these
are bits 8:0(of CPTX_AF_CONSTANTS0).

On 10x, Bit-8 of CPTX_AF_CONSTANTS0 is reserved and read
as zero. Let's use the common mask "0x1FF" to fetch total
numbers of "lf" on both 10x and 20x.

Change-Id: I572ef6ae3d129209674e2cc9e4e6824e7f1f05d3
Signed-off-by: Amit Singh Tomar <amitsinght@marvell.com>
Signed-off-by: Bharat Bhushan <bbhushan2@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/133629
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Integrated-by: Rui He <rui.he@windriver.com>