perf/marvell: Refactor to extract platform data - no functional change
commit 5c2c67a2fb538c430f4d3b3d4550ddcfec217113 from
git@git.assembla.com:cavium/WindRiver.linux.git

This commit introduces a refactor to the Marvell DDR pmu driver,
specifically targeting the extraction of platform data
(referred to as "pdata") from the existing driver. The purpose of
this refactor is to prepare for the upcoming support of the next
version of the Performance Monitoring Unit (PMU) in this
driver.

No functional changes are introduced in this refactor. Its sole
purpose is to improve code organization and pave the way for
future enhancements to the driver.

Signed-off-by: Gowthami Thiagarajan <gthiagarajan@marvell.com>
Change-Id: I6d9ed8eb804e0b0a3085ffe59f39e10f9dffa346
Signed-off-by: Siyu Zhang <siyu.zhang@windriver.com>
Reviewed-by: Zhaolong Zhang <zhaolong.zhang@windriver.com>