octeontx2-af: Add packet path between representor and VF
commit 2e37874218d45b3873deb3ae193b74aab9584b62 from
git@git.assembla.com:cavium/WindRiver.linux.git

Current HW, do not support in-built switch which will forward pkts
between representee and representor. When representor is put under
a bridge and pkts needs to be sent to representee, then pkts from
representor are sent on a HW internal loopback channel, which again
will be punted to ingress pkt parser. Now the rules that this patch
installs are the MCAM filters/rules which will match against these
pkts and forward them to representee.
The rules that this patch installs are for basic
representor <=> representee path similar to Tun/TAP between VM and
Host.

Signed-off-by: Geetha sowjanya <gakula@marvell.com>
Reviewed-by: Simon Horman <horms@kernel.org>
Change-Id: Ib050730d052d98b36ad6fe9626a398a078ef8cf8
Integrated-by: Rui He <rui.he@windriver.com>